divide scripts for controller page
