+++ src/routes/controller/scripts/config.js
@@ -1,0 +1,8 @@
+export let isDragging_conf = false;
+export let distanceJoystick_conf = 0;
+export let angleJoystick_conf = 0;
+export let batteryPercent_conf = 25;
+export let lives_conf = 3;
+export let joystickPosition_conf = { x: 0, y: 0};
+export let keysPressed_conf = { z: false, q: false, d: false, s: false};
+export const smoothingFactor = 0.1;
